% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_data.R
\name{flagp_update}
\alias{flagp_update}
\title{update FlaGP model}
\usage{
flagp_update(model, Xnew_orig, Tnew_orig = NULL, Ynew_orig, refit = F)
}
\arguments{
\item{model}{FlaGP model object}

\item{Xnew_orig}{new X to add on original scale}

\item{Ynew_orig}{new Y to add on original scale}

\item{refit}{should the covariance parameters be estimated again?}
}
\description{
updates FlaGP data object with additional simulation data. Currently incomplete, if refitting, needs to get all flags from previous model
}
\details{
Returns FlaGP data object
}
\examples{
# See examples folder for R markdown notebooks.

}
