% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_diagnostics.R
\name{plot.mcmc}
\alias{plot.mcmc}
\title{MCMC Plot}
\usage{
\method{plot}{mcmc}(x, labels = NULL, nrow = 2, ncol = 2, ...)
}
\arguments{
\item{x}{an \code{mcmc} object.}

\item{labels}{an optional vector of labels for each calibration parameter}

\item{nrow}{number of rows in display grid of plots}

\item{ncol}{number of columns in display grid of plots}

\item{...}{additional graphical parameters.}
}
\description{
Generate plots to assess MCMC
}
\details{
A pairs plot and trace plots are displayed in an nrow x ncol grid
}
\examples{
# See examples folder for R markdown notebooks.

}
