% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_data.R
\name{print.flagp}
\alias{print.flagp}
\title{print FlaGP model information}
\usage{
\method{print}{flagp}(flagp)
}
\arguments{
\item{flagp}{an \code{flagp} object.}
}
\description{
Prints relevant model information
}
