% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_seq_design.R
\name{seq_design}
\alias{seq_design}
\title{FlaGP sequential design}
\usage{
seq_design(
  model,
  n_cand = 100,
  n_int = 100,
  Xcand = NULL,
  Xint = NULL,
  seed = NULL,
  end = 50,
  n.pc = -1
)
}
\arguments{
\item{model}{FlaGP model to be augmented with a new simulation point}

\item{n_cand}{number of candidate points}

\item{n_int}{number of integration points}

\item{Xcand}{optional matrix of candidate points bounded between 0 and 1}

\item{Xint}{optional matrix of integration points bounded between 0 and 1}

\item{seed}{seed for reproducibility}

\item{end}{number of neighborhood points for prediction}

\item{n.pc}{number of PC's to use for design. Less can be much faster}
}
\description{
selects the X which reduces IMSE over the entire domain when added to the training set
}
\details{
Returns new X point
}
\examples{
# See examples folder for R markdown notebooks.

}
