% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_mcmc.R
\name{update_tuning_mv}
\alias{update_tuning_mv}
\title{this function updates a block Gaussian random walk proposal Metropolis-Hastings tuning parameters}
\usage{
update_tuning_mv(k, accept, lambda, batch_samples, Sigma_tune)
}
\arguments{
\item{k}{is a positive integer that is the current MCMC iteration.}

\item{accept}{is a positive number between 0 and 1 that represents the batch accpetance rate. The target of this adaptive tuning algorithm is an acceptance rate of between 0.44 (a univariate update) and 0.234 (5+ dimensional upate).}

\item{lambda}{is a positive scalar that scales the covariance matrix Sigma_tune and diminishes towards 0 as k increases.}

\item{batch_samples}{is a \eqn{batch.size \times d}{batch.size x d} dimensional matrix that consists of the batch samples of the d-dimensional parameter being sampled.}

\item{Sigma_tune}{is a \eqn{d \times d}{d x d} positive definite covariance matrix of the batch samples used to generate the multivariate Gaussian random walk proposal.}
}
\description{
this function updates a block Gaussian random walk proposal Metropolis-Hastings tuning parameters
}
\keyword{internal}
