% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_diagnostics.R
\name{contour_pairs}
\alias{contour_pairs}
\title{Plot MCMC posteriors}
\usage{
contour_pairs(
  theta.list,
  map = NULL,
  C.min = 0,
  C.max = 1,
  g.min = 0,
  g.max = 1,
  cols = c("cornflowerblue", "darkorange", "forestgreen"),
  labels = c("FlaGP")
)
}
\arguments{
\item{theta.list}{list of matrices containing posterior samples}

\item{map}{output from map estimation}

\item{cols}{colors for posterior density estimate plots}

\item{labels}{legend names for posterior density estimates}

\item{truth}{optional vector of known parameter values}
}
\description{
Pairs plot of posterior density estimates and prior distributions. This function is currently only implemented for a specific ball drop function. In future development it will be generalized.
}
