% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_data.R
\name{flagp}
\alias{flagp}
\title{FlaGP data object constructor}
\usage{
flagp(
  X.sim = NULL,
  T.sim = NULL,
  X.obs = NULL,
  T.obs = NULL,
  Y.sim,
  y.ind.sim = NULL,
  Y.obs = NULL,
  y.ind.obs = NULL,
  center = T,
  scale = T,
  scaletype = "scalar",
  n.pc = NULL,
  pct.var = 0.95,
  B = NULL,
  V.t = NULL,
  sigma.y = NULL,
  ls.subsample = "strat",
  ls.nugget = 1e-07,
  ls.m = 1,
  ls.K = 1,
  ls.prior = T,
  ls.parallel = T,
  make.cluster = T,
  ls.subsample.size = 250,
  bias = F,
  D = NULL,
  small = F,
  seed = NULL,
  verbose = T,
  rsvd = F
)
}
\arguments{
\item{X.sim}{experimental inputs for simulator output data}

\item{T.sim}{calibration inputs for simulator output data}

\item{X.obs}{observed inputs for experimental data}

\item{T.obs}{optional "known" calibration parameters for observations. This parameter does nothing in the model, but the data object will store scaled copies of it so it can be compared to calibration results}

\item{Y.sim}{simulator output matrix}

\item{y.ind.sim}{data indices for functional response simulations}

\item{Y.obs}{observed data matrix}

\item{y.ind.obs}{data indices for functional response observations}

\item{center}{center simulations to mean zero}

\item{scale}{scale simulations to unit variance}

\item{scaletype}{"scalar" or "rowwise" scaling to unit variance. Functional outputs should likely use "scalar" and multivariate outputs with different units should use "rowwise"}

\item{n.pc}{number of basis components to use for emulation. Defaults to 95\% variance explained.}

\item{pct.var}{choose number of basis components s.t. this proportion of variation is accounted for in simulations. Defaults to 95\% variance explained.}

\item{B}{optional matrix of basis vectors, if null B is computed via SVD(Y.sim)}

\item{V.t}{optional precomputed matrix of simulation weights for B}

\item{sigma.y}{assumed standard error of observations (not currently implemented)}

\item{ls.subsample}{type of subsampling procedure for lengthscale estimation. default 'blhs' is bootstrapped block latin hypercube sampling, 'strat' is stratified random sampling, and 'rand' is random sampling. Otherwise no subsampling is used.}

\item{ls.nugget}{gp nugget for lengthscale estimation}

\item{ls.m}{m parameter in laGP::blhs controlling data blocking for LHS subsample. Larger m will result in faster lengthscale estimation but is more memory intensive.}

\item{ls.K}{Number of bootstrap replicate estimations to do when ls.subsample is 'blhs','strat', or 'rand'. Estimates can be highly variable for small K.}

\item{ls.prior}{use detault prior in laGP::newGP for MAP lengthscale estimation, if FALSE, MLE estimation}

\item{ls.parallel}{do lengthscale estimation in parallel}

\item{ls.subsample.size}{size of subsamples when ls.subsample is 'strat' or 'rand'}

\item{bias}{calibration with a discrepancy model}

\item{D}{matrix of basis vectors for discrepancy model}

\item{small}{return small (memory) data object}

\item{verbose}{print status updates while building data object and doing precomputing}

\item{rsvd}{use random svd for fast svd calculations on massive Y.sim matrices. Must specify n.pc.}
}
\description{
Builds FlaGP data object and does necessary precomputing
}
\details{
Returns FlaGP data object
}
\examples{
# See examples folder for R markdown notebooks.

}
