% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_data.R
\name{flagp}
\alias{flagp}
\title{FlaGP data object constructor}
\usage{
flagp(
  X.sim = NULL,
  T.sim = NULL,
  X.obs = NULL,
  T.obs = NULL,
  Y.sim,
  y.ind.sim = NULL,
  Y.obs = NULL,
  y.ind.obs = NULL,
  center = T,
  scaletype = "scalar",
  n.pc = 1,
  pct.var = NULL,
  B = NULL,
  sigma.y = NULL,
  sc.nugget = 1e-07,
  sc.subsample = F,
  sc.subsample.size = 250,
  ls.prior = T,
  bias = F,
  D = NULL,
  small = F,
  precomp = T,
  verbose = F
)
}
\arguments{
\item{X.sim}{experimental inputs for simulator output data}

\item{T.sim}{calibration inputs for simulator output data}
}
\description{
Builds FlaGP data object and does necessary precomputing
}
\details{
Returns FlaGP data object
}
\examples{
# See examples folder for R markdown notebooks.

}
