% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_MAP.R
\name{map}
\alias{map}
\title{FlaGP MAP calibration}
\usage{
map(
  flagp,
  n.restarts = 1,
  init = NULL,
  seed = 1,
  opts = list(MAX_BB_EVAL = 10000, INITIAL_MESH_SIZE = 0.1, MIN_POLL_SIZE = "r0.001",
    DISPLAY_DEGREE = 0),
  end.eta = 50,
  delta.method = "newGP",
  start.delta = 6,
  end.delta = 50,
  theta.prior = "beta",
  theta.prior.params = c(2, 2),
  ssq.prior = "hcauchy",
  ssq.prior.params = c(0.1)
)
}
\arguments{
\item{flagp}{x an \code{flagp} object.}

\item{n.restarts}{size of random sample used to initialize the optimizer}

\item{init}{matrix of locations to initialize the optimizer}

\item{seed}{seed for reproducable initial conditions for the optimizer}

\item{opts}{list of options for crs::snomadr}
}
\description{
MAP estimate of calibration parameters
}
\details{
Returns FlaGP data object
}
\examples{
# See examples folder for R markdown notebooks.

}
