% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_mcmc.R
\name{mcmc}
\alias{mcmc}
\title{FlaGP MCMC with joint proposal}
\usage{
mcmc(
  flagp,
  t.init = rep(0.5, flagp$XT.data$p.t),
  ssq.init = 0.01,
  prop.cov = diag(c(rep(0.005, flagp$XT.data$p.t), 1e-06), flagp$XT.data$p.t + 1),
  n.samples = 10000,
  n.burn = 1000,
  adapt.par = c(100, 50, 0.5, n.burn),
  end.eta = 50,
  delta.method = "newGP",
  start.delta = 6,
  end.delta = 50,
  theta.prior = "beta",
  theta.prior.params = c(2, 2),
  ssq.prior = "hcauchy",
  ssq.prior.params = c(0.1),
  prev.samples = NULL,
  recalc.llh = F,
  sample = as.logical(ifelse(flagp$bias, F, T)),
  verbose = T
)
}
\arguments{
\item{flagp}{an \code{flagp} object.}

\item{t.init}{vector an inital value of the calibration parameters}

\item{n.samples}{integer number of mcmc samples to keep}

\item{n.burn}{integer number of initial samples to discard}

\item{adapt.par}{vector of parameters for addaptive metropolis. The first parameter it the first iteration to begin adaptation, the second is the frequency of adaptation, the third is the proportion of past samples to include for adaptation, and the fourth is the iteration to stop adaptation.}

\item{end.eta}{integer number of neighbors to use for laGP emulator}

\item{delta.method}{Model used for discrepancy. 'newGP' is a full GP using laGP default emperical bayes. 'rgasp' is a default RobustGaSP implementation and 'lagp' is an laGP predictor using the ALC criterion}

\item{start.delta}{initial neighborhood size for bias model. Only applicable if delta.method='lagp'}

\item{end.delta}{final neighborhood size for bias model. Only applicable if delta.method='lagp'}

\item{theta.prior}{prior for calibration parameters, 'beta' or 'unif'}

\item{theta.prior.params}{parameters for beta prior. Only applicable of theta.prior='beta'}

\item{ssq.prior.params}{gamma prior parameters for error variance facilitating conjugate updating}

\item{prev.samples}{Use in accordance with add_samples(). A matrix of samples can be passed in to be used for proposal covariance estimation.}

\item{recalc.llh}{Should the likelihood be recalculated at every iteration. Generally this should be False and only considered if sample=T. Recalculing the likelihood with a new sample from the emulator (and discrepancy) can be useful if mcmc mixing is a problem. This is only likely to happen if a discrepancy model is being used or if the emulator poorly fits the observations and no discrepancy model is used. Importantly, inference is approximate if recalculation is used, as posterior variance will be artificially increased.}

\item{sample}{Account for uncertainty in the emulator and discrepancy model by sampling rather than by including the variance in the likelihood. Default is sample=T when bias=F which works well because sampling variability in the emulator is usually small. For a biased calibration we recommend sample=F. This will increase the cost of likelihood calculation, but result in a much more stable mcmc.}

\item{verbose}{print status updates}

\item{prop.sd.init}{vector of initial proposal distribution sd's}
}
\description{
Addaptive MCMC as defined in Haario et al. 2001 - "An adaptive Metropolis algorithm"
}
\details{
Returns predictions at X.pred.orig
}
\examples{
# See examples folder for R markdown notebooks.

}
