% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_mcmc.R
\name{mcmc}
\alias{mcmc}
\title{FlaGP MCMC with joint proposal}
\usage{
mcmc(
  flagp,
  t.init = NULL,
  cov.init = NULL,
  n.samples = 10000,
  n.burn = 1000,
  update.every = 1,
  stop.update = 1000,
  end.eta = 50,
  lagp.delta = F,
  start.delta = 6,
  end.delta = 50,
  theta.prior = "beta",
  ssq.prior = "gamma",
  prev.samples = NULL,
  verbose = T
)
}
\arguments{
\item{flagp}{an \code{flagp} object.}

\item{t.init}{vector an inital value of the calibration parameters}

\item{cov.init}{matrix initial covariance matrix for the proposal distribution}

\item{n.samples}{integer number of mcmc samples to keep}

\item{n.burn}{integer number of initial samples to discard}

\item{update.every}{integer sample frequency that proposal covariance should be updated}

\item{stop.update}{integer sample index where proposal covariance updates should stop}

\item{end.eta}{integer number of neighbors to use for laGP emulator}

\item{lagp.delta}{T/F laGP bias model. If F, a full GP is used}

\item{start.delta}{initial neighborhood size for bias model. Only applicable if lagp.delta=T}

\item{end.delta}{final neighborhood size for bias model. Only applicable if lagp.delta=T}

\item{theta.prior}{prior for calibration parameters, 'beta' or 'unif'}

\item{ssq.prior}{prior for error variance, 'gamma' is a non-informative gamma on the precision}

\item{prev.samples}{Use in accordance with add_samples(). A matrix of samples can be passed in to be used for proposal covariance estimation.}

\item{verbose}{print status updates}
}
\description{
Addaptive MCMC as defined in Haario et al. 2001 - "An adaptive Metropolis algorithm"
}
\details{
Returns predictions at X.pred.orig
}
\examples{
# See examples folder for R markdown notebooks.

}
