% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_mcmc.R
\name{parallel_mcmc}
\alias{parallel_mcmc}
\title{Parallel wrapper for mcmc function}
\usage{
parallel_mcmc(
  flagp,
  t.init = rep(0.5, flagp$XT.data$p.t),
  prop.cov = diag((0.5/3)^2, flagp$XT.data$p.t),
  n.samples = 10000,
  n.burn = 1000,
  update.every = 50,
  stop.update = 1000,
  end.eta = 50,
  delta.method = "newGP",
  start.delta = 6,
  end.delta = 50,
  theta.prior = "beta",
  theta.prior.params = c(2, 2),
  ssq.prior.params = c(1, 0.001),
  prev.samples = NULL,
  recalc.llh = F,
  prop.type = "mv",
  verbose = T,
  n.chains = 2,
  n.cores = 2
)
}
\arguments{
\item{flagp}{an \code{flagp} object.}

\item{t.init}{vector an inital value of the calibration parameters}

\item{n.samples}{integer number of mcmc samples to keep}

\item{n.burn}{integer number of initial samples to discard}

\item{update.every}{integer sample frequency that proposal covariance should be updated, only working for update.every=1 currently}

\item{stop.update}{integer sample index where proposal covariance updates should stop}

\item{end.eta}{integer number of neighbors to use for laGP emulator}

\item{delta.method}{Model used for discrepancy. 'newGP' is a full GP using laGP default emperical bayes. 'rgasp' is a default RobustGaSP implementation and 'lagp' is an laGP predictor using the ALC criterion}

\item{start.delta}{initial neighborhood size for bias model. Only applicable if delta.method='lagp'}

\item{end.delta}{final neighborhood size for bias model. Only applicable if delta.method='lagp'}

\item{theta.prior}{prior for calibration parameters, 'beta' or 'unif'}

\item{ssq.prior.params}{gamma prior parameters for error variance facilitating conjugate updating}

\item{prev.samples}{Use in accordance with add_samples(). A matrix of samples can be passed in to be used for proposal covariance estimation.}

\item{verbose}{print status updates}

\item{n.chains}{number of parallel chains to run}

\item{n.cores}{number of cores for parallel processing}

\item{prop.sd.init}{matrix initial covariance matrix for the proposal distribution}
}
\description{
Addaptive MCMC as defined in Haario et al. 2001 - "An adaptive Metropolis algorithm"
}
\details{
Returns predictions at X.pred.orig
}
\examples{
# See examples folder for R markdown notebooks.

}
