% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_diagnostics.R
\name{plot.flagp}
\alias{plot.flagp}
\title{FlaGP data plots}
\usage{
\method{plot}{flagp}(flagp, basis = T, legend = T, xlab = "x", ylab = "y", ...)
}
\arguments{
\item{flagp}{an \code{flagp} object.}

\item{basis}{T/F indicating if the bases should be plotted}

\item{legend}{T/F indicating if legends should be added to the plots}

\item{xlab}{an optional descriptive label for the data x axes}

\item{ylab}{an optional descriptive label for the data y axes}

\item{...}{additional graphical parameters.}
}
\description{
Plots X,Y data as well as computed bases
}
\details{
Plots the X,Y data, both simulated and observed and optionally plots the bases as well
}
\examples{
# See examples folder for R markdown notebooks.

}
