% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlaGP_predict.R
\name{predict.flagp}
\alias{predict.flagp}
\title{FlaGP Prediction}
\usage{
\method{predict}{flagp}(
  flagp,
  model,
  X.pred.orig,
  n.samples = 1,
  samp.ids = NULL,
  return.samples = F,
  support = "obs",
  end.eta = 50,
  lagp.delta = F,
  start.delta = 6,
  end.delta = 50
)
}
\arguments{
\item{model}{an \code{mcmc} or \code{map} object.}
}
\description{
Prediction with \code{mcmc} or \code{map} object
}
\details{
Returns predictions at X.pred.orig
}
\examples{
# See examples folder for R markdown notebooks.

}
